backend/create_recipe/requirements.txt
PyJWT==2.8.0
boto3==1.34.144
requests==2.32.3
urllib3==2.2.2
openai==1.35.13
DateTime==5.5